Platform separation (#4178)
* Added system dark mode support for primary label colors in the framework

Moved Color to its own Platform file

* Relocated color file

* Reverted demo change

* Separated `Platform` into multiple files for better organization

* Updated from master
