+++ ChartsDemo-iOS/Swift/Demos/PiePolylineChartViewController.swift
@@ -6,9 +6,7 @@
 //  Copyright © 2017 jc. All rights reserved.
 //
 
-#if canImport(UIKit)
-    import UIKit
-#endif
+import UIKit
 import Charts
 
 class PiePolylineChartViewController: DemoBaseViewController {
