+++ Source/Charts/Charts/ChartViewBase.swift
@@ -98,7 +98,7 @@ open class ChartViewBase: NSUIView, ChartDataProvider, AnimatorDelegate
     @objc open var noDataFont = NSUIFont.systemFont(ofSize: 12)
     
     /// color of the no data text
-    @objc open var noDataTextColor: NSUIColor = NSUIColor.labelOrBlack
+    @objc open var noDataTextColor: NSUIColor = .labelOrBlack
 
     /// alignment of the no data text
     @objc open var noDataTextAlignment: NSTextAlignment = .left
